[BUILD] don't discard symbols
Symbols are needed to report useful backtraces.
And debug symbols are more or less required to
debug.

Using --disable-stripping will ensure the symbols
generated by the compiler, depending on the CFLAGS,
will be kept, at least in the staging directory.

Anyway, they will be removed by default when the
library is embedded in a firmware.