[DEV] atom.mk: support USE_CLANG=1 build
When building for a native linux target with USE_CLANG=1, ffmpeg keep
trying to build with GCC instand of clang, and ./configure fail with
the following error message:

  Autotools: ffmpeg-libav => Configuring
  gcc is unable to create an executable file.
  C compiler test failed.

  If you think configure made a mistake, make sure you are using the latest
  version from Git.  If the latest version fails, report the problem to the
  ffmpeg-user@ffmpeg.org mailing list or IRC #ffmpeg on irc.freenode.net.
  Include the log file "ffbuild/config.log" produced by configure as this will help
  solve the problem.

And config.log last lines are:

  gcc -isystem[...] -std=gnu99 -pipe -O2 -g -ffunction-sections -DALCHEMY_BUILD -m64 -fPIC -msse -msse2 -mssse3 -Qunused-arguments -fpch-preprocess -include [...] -DBUILD_ZLIB -c -o /tmp/ffconf.kz7mrR8o/test.o /tmp/ffconf.kz7mrR8o/test.c
  gcc: error: unrecognized command-line option '-Qunused-arguments'

It seems ffmpeg's ./configure doesn't make use of CC or CXX environment
variables. Instead, it requires usage of --cc or --cxx parameter.

With this patch, the full path of the compiler (clang, gcc) is passed
to ./configure: it will be used regardless of the cross-compiler
prefix.